Add functionality to distributed manager to provide mesh-wide groups.
A pytorch DeviceMesh provides syntax to to access the ProcessGroup across a mesh dimension.
Sometimes, it is most easy to access a group that includes all devices of a mesh.  This isn't
included in the upstream syntax, so it's built out here.  Because the cost of creating a group
is not small, the groups are cached using the devicemesh itself (hashed) as a key.

In 1D meshes, the underlying group for that mesh is returned rather than creating a new group.