Reorganize the halo and ring message passing to be easier to follow and maintain.
This reorganization isolates the layers of halo passing by conceptual
operation (building the halos, communicate halos, apply halos, slice
off residuals, etc).  The code is a little longer but the upshot
is the availability of high level functions `halo_padding` and `unhalo_padding`
which are both differentiable and easily applied.

Also introduces a ring message passing function.  Note that this function
is synchronous (so is the halo) and while the halo needs to be synchronous,
ring message passing often does not.  It's included nevertheless as a
simple, easy to use version to enable debugging of the overlapped version.

Both Halos and Rings and now configured with light dataclass objects
to make the number of arguments passed around simpler, and easier to maintain
state between forward and backward passes.